+++ CHANGELOG.md
@@ -2,7 +2,7 @@
   * [Fixed] #722 Fix regression with addChild performance improvements (#708)
 
 # BRANCH aria-grid / Unreleased
-  * 
+  * [Added] `ext-ariagrid` (experimental)
 
 # 2.22.0 / 2017-04-11
   * [Added] ext-dnd5 now part of standard distribution